search-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
supabase==2.3.0
httpx==0.26.0
python-dotenv==1.0.0
rapidfuzz==3.6.1
sse-starlette==1.8.2
pytest==7.4.3
pytest-asyncio==0.23.3
python-jose[cryptography]==3.3.0
